Move plot_pdf to vb_visu.py
